+++ test/slice_test.cc
@@ -4,8 +4,8 @@
 
 int main(int argc, char *argv[])
 {
-    std::vector<int> vec;
-    std::vector<std::vector<int>> vecc;
+    std::vector<uint16_t> vec;
+    std::vector<std::vector<uint16_t>> vecc;
     for (int i = 0; i < 234; i++) {
         vec.push_back(i);
     }
